add binary search algorithm
Basic implementation of binary search in Python. The script already uses
a list as example so by running `python binary_search.py` it returns the
position of the element in the example list.

This is a pretty straight forward algorithm but it was implemented while
studying the book "Grokking Algorithms" by Aditya Bhargava.